requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.25.0
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.49.0
django-ckeditor==6.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==2.1.0
gunicorn==20.0.4
idna==3.3
oauthlib==3.2.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
